#33 wip split api directory
