+++ src/features/currency/api/client.ts
@@ -1,20 +1,20 @@
-import { PIVOT_CURRENCY_CODE } from '../../constants/currency';
-import type { Results } from '../../types/common';
+import { ApiCache } from '../../../api/cache';
+import { fetchString } from '../../../api/client';
+import { PIVOT_CURRENCY_CODE } from '../../../constants/currency';
+import type { Results } from '../../../types/common';
 import type {
   Currency,
   DateRate,
   CodeString,
   SymbolString,
-} from '../../types/currency';
+} from '../../../types/currency';
 import {
   parseCsv,
   createPivotCurrency,
   createCurrency,
   splitSymbolToCodes,
   createCrossCurrency,
-} from '../../utils/currency';
-import { ApiCache } from '../common/cache';
-import { fetchString } from '../common/client';
+} from '../../../utils/currency';
 
 const cache = new ApiCache<Currency>();
